+++ src/main/java/com/example/fms/festmanagement/controller/UserDashboardController.java
@@ -1,5 +1,6 @@
 package com.example.fms.festmanagement.controller;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,6 +22,7 @@ import com.example.fms.festmanagement.models.Item;
 import com.example.fms.festmanagement.models.Participant;
 import com.example.fms.festmanagement.models.Participation;
 import com.example.fms.festmanagement.models.SubEvent;
+import com.example.fms.festmanagement.models.Transaction;
 import com.example.fms.festmanagement.models.User;
 import com.example.fms.festmanagement.service.AuthenticationService;
 import com.example.fms.festmanagement.service.DashboardService;
@@ -43,11 +45,14 @@ public class UserDashboardController extends Helper {
     private MessageService messageService;
 
     @GetMapping("/dashboard")
-    public String dashboard(Model model, HttpSession session) {
+    public String dashboard(Model model, HttpSession session, RedirectAttributes attributes) {
         if (!isAuthenticated(session)) {
             return "redirect:/";
         }
         addDefaultAttributes(model, session);
+        if(!model.getAttribute("userRole").equals("participant")){
+            return "redirect:accessdenied";
+        }
         model.addAttribute("subEvents", dashboardService.getSubEventToday());
         model.addAttribute("correspondingevents", dashboardService.getEventFromSubEvent(dashboardService.getSubEventToday()));        
         return "dashboard";
@@ -55,6 +60,13 @@ public class UserDashboardController extends Helper {
 
     @GetMapping("showevents")
     public String ShowEvents(Model model, HttpSession session){
+        if (!isAuthenticated(session)) {
+            return "redirect:/";
+        }
+        addDefaultAttributes(model, session);
+        if(!model.getAttribute("userRole").equals("participant")){
+            return "redirect:accessdenied";
+        }
         List<Event> e=dashboardService.showEvents();
         List<List<SubEvent>>s=dashboardService.showSubEvents(e);
         List<List<List<Competition>>>c=dashboardService.showCompetitions(s);
@@ -68,6 +80,13 @@ public class UserDashboardController extends Helper {
 
     @GetMapping("{eventId}/{subEventId}/{competitionId}/participate")
     public String Participate(@PathVariable("eventId") int eventId, @PathVariable("subEventId") int subEventId, @PathVariable("competitionId") int competitionId, Model model, HttpSession session, RedirectAttributes attributes){
+        if (!isAuthenticated(session)) {
+            return "redirect:/";
+        }
+        addDefaultAttributes(model, session);
+        if(!model.getAttribute("userRole").equals("participant")){
+            return "redirect:accessdenied";
+        }
         Participation p=new Participation();
         addDefaultAttributes(model, session);
         p.setParticipantEmail(authenticationService.getCurrentUser(session));
@@ -83,6 +102,13 @@ public class UserDashboardController extends Helper {
 
     @GetMapping("{eventId}_{subEventId}_{competitionId}_leaderboard")
     public String Leaderboard(@PathVariable("eventId") int eventId, @PathVariable("subEventId") int subEventId, @PathVariable("competitionId") int competitionId, Model model, HttpSession session){
+        if (!isAuthenticated(session)) {
+            return "redirect:/";
+        }
+        addDefaultAttributes(model, session);
+        if(!model.getAttribute("userRole").equals("participant")){
+            return "redirect:accessdenied";
+        }
         Event e= organiserDashboardService.getEvent(eventId);
         SubEvent s = organiserDashboardService.getSubEventById(subEventId,e);
         Competition c =organiserDashboardService.getCompetitionById(competitionId,s);
@@ -100,6 +126,13 @@ public class UserDashboardController extends Helper {
 
     @GetMapping("goodies")
     public String Goodies(Model model, HttpSession session){
+        if (!isAuthenticated(session)) {
+            return "redirect:/";
+        }
+        addDefaultAttributes(model, session);
+        if(!model.getAttribute("userRole").equals("participant")){
+            return "redirect:accessdenied";
+        }
         Cart c=dashboardService.getActiveCart(authenticationService.getCurrentUser(session));
         if(c==null){
             dashboardService.makeCart(authenticationService.getCurrentUser(session));
@@ -114,6 +147,13 @@ public class UserDashboardController extends Helper {
 
     @GetMapping("cart")
     public String Cart(Model model, HttpSession session){
+        if (!isAuthenticated(session)) {
+            return "redirect:/";
+        }
+        addDefaultAttributes(model, session);
+        if(!model.getAttribute("userRole").equals("participant")){
+            return "redirect:accessdenied";
+        }
         Cart c=dashboardService.getActiveCart(authenticationService.getCurrentUser(session));
         if(c==null){
             dashboardService.makeCart(authenticationService.getCurrentUser(session));
@@ -131,6 +171,13 @@ public class UserDashboardController extends Helper {
 
     @GetMapping("{itemId}/additem")
     public String AddGoodies(Model model, HttpSession session, @PathVariable("itemId") int itemId){
+        if (!isAuthenticated(session)) {
+            return "redirect:/";
+        }
+        addDefaultAttributes(model, session);
+        if(!model.getAttribute("userRole").equals("participant")){
+            return "redirect:accessdenied";
+        }
         Cart c=dashboardService.getActiveCart(authenticationService.getCurrentUser(session));
         
         if(c==null){
@@ -143,6 +190,13 @@ public class UserDashboardController extends Helper {
 
     @PostMapping("updateitemincart")
     public String UpdateGoodies(Model model, HttpSession session, @ModelAttribute("cid") CartItemDetails cid){
+        if (!isAuthenticated(session)) {
+            return "redirect:/";
+        }
+        addDefaultAttributes(model, session);
+        if(!model.getAttribute("userRole").equals("participant")){
+            return "redirect:accessdenied";
+        }
         System.out.println(cid.toString());
         dashboardService.updateCart(cid);
         return "redirect:/cart";
@@ -150,6 +204,13 @@ public class UserDashboardController extends Helper {
 
     @GetMapping("{itemId}/removeitem")
     public String RemoveGoodie(Model model, HttpSession session,@PathVariable("itemId") int itemId){
+        if (!isAuthenticated(session)) {
+            return "redirect:/";
+        }
+        addDefaultAttributes(model, session);
+        if(!model.getAttribute("userRole").equals("participant")){
+            return "redirect:accessdenied";
+        }
         Cart c=dashboardService.getActiveCart(authenticationService.getCurrentUser(session));
         dashboardService.removeFromCart(c,itemId);
         return "redirect:/cart";
@@ -157,13 +218,40 @@ public class UserDashboardController extends Helper {
 
     @GetMapping("previoustransactions")
     public String PreviousTransactions(Model model, HttpSession session){
+        if (!isAuthenticated(session)) {
+            return "redirect:/";
+        }
+        addDefaultAttributes(model, session);
+        if(!model.getAttribute("userRole").equals("participant")){
+            return "redirect:accessdenied";
+        }
         List<Cart> c=dashboardService.getPreviousCarts(authenticationService.getCurrentUser(session));
-        model.addAttribute("c",c);
+        List<Transaction> t = new ArrayList<Transaction>();
+        for(Cart cart:c){
+            t.add(dashboardService.getTransaction(cart));
+        }
+        List<List<Item>>ret=new ArrayList<List<Item>>();
+        List<List<CartItemDetails>>ret2=new ArrayList<List<CartItemDetails>>();
+        for(Cart cart:c){
+            List<CartItemDetails> cid =dashboardService.getCartItemDetails(cart);
+            ret2.add(cid);
+            ret.add(dashboardService.getItemDetails(cid));
+        }
+        model.addAttribute("transactions",t);
+        model.addAttribute("items",ret);
+        model.addAttribute("cid",ret2);
         return "previoustransactions";
     }
 
     @GetMapping("checkout")
     public String Checkout(Model model, HttpSession session){
+        if (!isAuthenticated(session)) {
+            return "redirect:/";
+        }
+        addDefaultAttributes(model, session);
+        if(!model.getAttribute("userRole").equals("participant")){
+            return "redirect:accessdenied";
+        }
         model.addAttribute("user",dashboardService.getParticipant(authenticationService.getCurrentUser(session)));
         Cart c=dashboardService.getActiveCart(authenticationService.getCurrentUser(session));
         List<CartItemDetails>cid=dashboardService.getCartItemDetails(c);
@@ -179,6 +267,13 @@ public class UserDashboardController extends Helper {
 
     @PostMapping("checkout")
     public String PostCheckout(Model model, HttpSession session, @ModelAttribute Participant p){
+        if (!isAuthenticated(session)) {
+            return "redirect:/";
+        }
+        addDefaultAttributes(model, session);
+        if(!model.getAttribute("userRole").equals("participant")){
+            return "redirect:accessdenied";
+        }
         Cart c=dashboardService.getActiveCart(authenticationService.getCurrentUser(session));
         dashboardService.updateParticipant(p);
         dashboardService.createTransaction(c);
